% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateInitialValues.R
\name{CalculateInitialValues}
\alias{CalculateInitialValues}
\title{Title}
\usage{
CalculateInitialValues(f, split_result = split_result)
}
\arguments{
\item{f}{a numerical array}

\item{split_result}{a numerical vector, to split indexes of array}
}
\value{
a numerical vector
}
\description{
Title
}
\examples{
set.seed(1234)
s1 <- sample(c(1,2), size=500, replace=TRUE)
s2 <- sample(c(1,2), size=500, replace=TRUE)
f <- ProbMatrixF(cbind(s1,s2))
CalculateInitialValues(f, c(1,2)) ##for first equation
CalculateInitialValues(f, c(3,4)) ##for second equation
}
