% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{Inference}
\alias{Inference}
\title{Title}
\usage{
Inference(ni, qi, lambda)
}
\arguments{
\item{ni}{numerical matrix of transitions frequencies}

\item{qi}{numerical matrix of transitions probabilities}

\item{lambda}{numerical vector}
}
\value{
list with standard error, z-statistics and pvalues associated with each estimate
}
\description{
Title
}
\examples{
#########
}
