% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference_x.R
\name{Inference_x}
\alias{Inference_x}
\title{Title}
\usage{
Inference_x(hess, lambda)
}
\arguments{
\item{hess}{numerical matrix}

\item{lambda}{numerical vector}
}
\value{
list with standard error, z-statistics and pvalues associated with each estimate
}
\description{
Title
}
\examples{
######
}
