% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table.R
\name{output.table}
\alias{output.table}
\title{Title}
\usage{
output.table(estimates, se, zstat, pvalue, ll, eq, flag = 0)
}
\arguments{
\item{estimates}{numerical vector of estimates}

\item{se}{numerical vector of standard error of estimates}

\item{zstat}{numerical vector of z-statistics of estimates}

\item{pvalue}{numerical vector of pvalue}

\item{ll}{numerical value of loglikelihood value}

\item{eq}{numerical value of number of equation}

\item{flag}{numerical value of warning in optimization}
}
\value{
prints the results table
}
\description{
Title
}
\examples{
#####


}
