% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialDerivatives.R
\name{PartialDerivatives}
\alias{PartialDerivatives}
\title{Title}
\usage{
PartialDerivatives(ni, qi, lambda)
}
\arguments{
\item{ni}{numerical matrix of number of transitions between states and data sequences}

\item{qi}{numerical matrix of transitions probabilities between states and data sequences}

\item{lambda}{numerical vector}
}
\value{
numerical vector with partial derivates values of each lambda
}
\description{
Title
}
\examples{
#######
}
