% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbMatrixF.R
\name{ProbMatrixF}
\alias{ProbMatrixF}
\title{Title}
\usage{
ProbMatrixF(s)
}
\arguments{
\item{s}{numerical matrix with categorical data sequences}
}
\value{
numerical array with frequency transition matrices
}
\description{
Title
}
\examples{
##############
}
