% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbMatrixQ.R
\name{ProbMatrixQ}
\alias{ProbMatrixQ}
\title{Title}
\usage{
ProbMatrixQ(s)
}
\arguments{
\item{s}{numerical matrix with categorical data sequences}
}
\value{
numerical array with probability transition matrices
}
\description{
Title
}
\examples{
##############
}
