% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbValuesXDependent.R
\name{ProbValuesXDependent}
\alias{ProbValuesXDependent}
\title{Title}
\usage{
ProbValuesXDependent(s, x)
}
\arguments{
\item{s}{numerical matrix with categorical data sequences}

\item{x}{Matrix of covariates (exogeneous variables)}
}
\value{
non-homogeneous probability transition matrix
}
\description{
Title
}
\examples{
#############
}
