% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcx.R
\name{mmcx}
\alias{mmcx}
\title{Title}
\usage{
mmcx(y, x, initial, ...)
}
\arguments{
\item{y}{matrix of categorical data sequences}

\item{x}{matrix of covariates}

\item{initial}{numerical vector of initial values.}

\item{...}{additional arguments to be passed down to auglag()}
}
\value{
The function returns the parameter estimates, standard-errors, z-statistics, p-values and the value of the log-likelihood function, for each equation.
}
\description{
Title
}
\examples{
set.seed(1234)
s1 <- sample(c(1,2), 500, replace=TRUE)
s2 <- sample(c(1,2), 500, replace=TRUE)
x <- rnorm(500)
mmcx(y = cbind(s1,s2), x = cbind(x), initial=c(1,1))

}
