% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.mtd.R
\name{multi.mtd}
\alias{multi.mtd}
\title{Title}
\usage{
multi.mtd(y, deltaStop = 1e-04, is_constrained = TRUE, delta = 0.1)
}
\arguments{
\item{y}{matrix of categorical data sequences}

\item{deltaStop}{value below which the optimization phases of the parameters stop}

\item{is_constrained}{flag indicating wether the function will consider the usual set of constraints (usual set: TRUE, new set of constraints: FALSE).}

\item{delta}{the amount of change to increase/decrease in the parameters for each iteration of the optimization algorithm.}
}
\value{
The function returns the parameter estimates, standard-errors, z-statistics, p-values and the value of the log-likelihood function, for each equation.
}
\description{
Title
}
\examples{
set.seed(1234)
s1 <- sample(c(1,2), 500, replace=TRUE)
s2 <- sample(c(1,2), 500, replace=TRUE)
multi.mtd(y = cbind(s1,s2))

}
