% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.mtd_probit.R
\name{multi.mtd_probit}
\alias{multi.mtd_probit}
\title{Title}
\usage{
multi.mtd_probit(y, initial, nummethod = "bfgs")
}
\arguments{
\item{y}{matrix of categorical data sequences}

\item{initial}{numerical vector of initial values}

\item{nummethod}{Numerical maximisation method, currently either "NR" (for Newton-Raphson), "BFGS" (for Broyden-Fletcher-Goldfarb-Shanno), "BFGSR" (for the BFGS algorithm implemented in R), "BHHH" (for Berndt-Hall-Hall-Hausman), "SANN" (for Simulated ANNealing), "CG" (for Conjugate Gradients), or "NM" (for Nelder-Mead). Lower-case letters (such as "nr" for Newton-Raphson) are allowed. The default method is "BFGS". For more details see maxLik.}
}
\value{
The function returns the parameter estimates, standard-errors, z-statistics, p-values and the value of the log-likelihood function, for each equation.
}
\description{
Title
}
\examples{
set.seed(1234)
s1 <- sample(c(1,2), 500, replace=TRUE)
s2 <- sample(c(1,2), 500, replace=TRUE)
multi.mtd_probit(y = cbind(s1,s2), initial=c(1,1,1), nummethod='bfgs')

}
